Move all the unneeded imports out of __init__
We only strictly need __version__, but should "from cats import *"
give us anything else?
